Migrate Input handler to Environment
